+++ src/specific/s_music_mm.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "global/types.h"
-#include "specific/s_music.h"
-
-S_MUSIC_BACKEND *S_Music_MM_Factory(void);