+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - `/pos`
     - `/tp [room_num]`
     - `/tp [x] [y] [z]`
+- changed the music backend to SDL + libav
 
 ## [0.1.1](https://github.com/LostArtefacts/TR2X/compare/0.1...0.1.1) - 2024-04-27
 - fixed Lara's shadow with z-buffer option on (#64, regression from 0.1)